requirement.txt

# LangChain Core
langchain
langchain-core

#langchain_community
langchain_community

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# Utilities
pydantic
grandalf    # used to draw the flow of execution
pypdf
pdfplumber
pymupdf
beautifulsoup4
lxml


